requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.6
django-ckeditor==6.1.0
django-embed-video==1.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-utils-six==2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
jsmin==3.0.0
Pillow==8.3.2
psycopg2==2.9.3
python-decouple==3.5
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.3.0
